// src/app/interfaces/grid/igrid-page-options.ts
/**
 * @showTitle  'true' to show a title, 'false' to leave off title
 * @title  the optional title for the grid
 * @showRefreshIcon  'true' to show the icon, 'false' to leave it off the grid
 * @showResetIcon  'true' to show the icon, 'false' to leave it off the grid
 * @showPDFIcon  'true' to show the icon, 'false' to leave it off the grid
 * @showSaveIcon  'true' to show the icon, 'false' to leave it off the grid
 * @showAddIcon  'true' to show the icon, 'false' to leave it off the grid
 */
export interface IGridPageOptions {
  showTitle: boolean;
  title: string;
  showRefreshIcon: boolean;
  showResetIcon: boolean;
  showPDFIcon: boolean;
  showSaveIcon: boolean;
  showAddIcon: boolean;
}